new commit

+++ src/main/java/org/yearup/controllers/CategoriesController.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.yearup.data.CategoryDao;
 import org.yearup.data.ProductDao;
 import org.yearup.models.Category;
@@ -18,7 +19,7 @@ import java.util.List;
 
 
 @RestController
-@RequestMapping("/categories")
+@RequestMapping("categories")
 @CrossOrigin
 public class CategoriesController {
     private CategoryDao categoryDao;
@@ -32,57 +33,90 @@ public class CategoriesController {
     }
 
 
-    @GetMapping
+    @GetMapping("")
     public List<Category> getAll() {
 
         return categoryDao.getAllCategories();
     }
 
-    @GetMapping("/ {id}")
+    @GetMapping("/{id}")
     public Category getById(@PathVariable int id) {
+        Category category = categoryDao.getById(id);
+        if (category == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Oops, error occurred. Try again.");
+        }
+        return category;
+//        try {
+//            var category = categoryDao.getById(id);
+//            if (category == null)
+//                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+//            return category;
+//        } catch (ResponseStatusException ex) {
+//            throw ex;
+//        } catch (Exception ex) {
+//            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops, error occurred. Try again");
+//        }
 //        return categoryDao.getById(id);
-        return null;
+//        return null;
 
     }
 
-    // the url to return all products in category 1 would look like this
-    // https://localhost:8080/categories/1/products
+
     @GetMapping("/{categoryId}/products")
     public List<Product> getProductsById(@PathVariable int categoryId)
     {
-
         return productDao.listByCategoryId(categoryId);
         //return productDao.getBYId(id);
     }
 
     @PostMapping()
-    @PreAuthorize("hasRole( 'ADMIN' )")
-//    @RequestMapping(path = " /categories", method = RequestMethod.POST)
-    @ResponseStatus(value = HttpStatus.CREATED)
+    @PreAuthorize("hasRole('ADMIN')")
+    @ResponseStatus(HttpStatus.CREATED)
     public Category addCategory(@RequestBody Category category) {
-//        Product newProduct = productDao.insert(product);
-       return categoryDao.create(category);
+        return categoryDao.create(category);
+
+//        try {
+//            return categoryDao.create(category);
+//        } catch (Exception ex) {
+//            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops, error occurred. Try again.");
+//        }
+
 
     }
 
 
-    @PreAuthorize("hasRole ('ADMIN') ")
-    @PutMapping("/{id}")
-//    @RequestMapping(path = "/{id}", method = RequestMethod.PUT)
-//    @ResponseStatus(HttpStatus.NO_CONTENT)
+   @PutMapping("/{id}")
+   @PreAuthorize("hasRole('ADMIN')")
     public void updateCategory(@PathVariable int id, @RequestBody Category category) {
-//        category.setCategoryId(id);
         categoryDao.update(id, category);
 
+//        try {
+//            categoryDao.update(id, category);
+//        } catch (Exception e) {
+//            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops, error occurred. Try again.");
+//        }
+
+
     }
 
 
-    @PreAuthorize("hasRole ('ADMIN') ")
-    @DeleteMapping("/{id}")
-//    @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
-    @ResponseStatus(HttpStatus.NO_CONTENT)
+   @DeleteMapping("/{id}")
+   @PreAuthorize("hasRole('ADMIN')")
+   @ResponseStatus(value = HttpStatus.NO_CONTENT)
     public void deleteCategory(@PathVariable int id) {
+        var category = categoryDao.getById(id);
+        if (category == null)
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
         categoryDao.delete(id);
+//        try {
+//            var category = categoryDao.getById(id);
+//            if (category == null)
+//                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+//            categoryDao.delete(id);
+//        } catch (Exception ex) {
+//            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops, error occurred. Try again.");
+//        }
